feat: 친구 팔로우, 언팔로우, 내점수, 뱃지, 스트릭

+++ pocketc_android/app/src/main/java/com/example/irumi/data/dto/response/main/DailySavingResponse.kt
@@ -7,5 +7,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DailySavingResponse(
     @SerialName("savingScore") val savingScore: Int,
-    @SerialName("totalSpending") val totalSpending: Long
+    @SerialName("totalSpending") val totalSpending: Int
 )
