+++ pocketc_android/app/src/main/java/com/example/irumi/data/dto/response/main/UserProfileResponse.kt
@@ -8,6 +8,6 @@ import kotlinx.serialization.Serializable
 data class UserProfileResponse(
     @SerialName("userId") val userId: Int,
     @SerialName("name") val name: String,
-    @SerialName("budget") val budget: Long, // 서버 숫자 타입에 맞춰 Int/Long 결정
+    @SerialName("budget") val budget: Int, // 서버 숫자 타입에 맞춰 Int/Long 결정
     @SerialName("profileImageUrl") val profileImageUrl: String
 )
